requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
h11==0.12.0
idna==3.2
ifaddr==0.1.7
paho-mqtt==1.5.1
pickleDB==0.9.2
pycparser==2.20
requests==2.26.0
toml==0.10.2
urllib3==1.26.6
zeroconf==0.35.0
